+++ src/main/java/absolutelyaya/ultracraft/item/AbstractRevolverItem.java
@@ -68,10 +68,15 @@ public abstract class AbstractRevolverItem extends AbstractWeaponItem implements
 			}
 			
 			if(isAlternate())
-				ServerHitscanHandler.makeBasicHitscan(user, ServerHitscanHandler.SLAB, 2 * getPrimaryDamage(), DamageSources.GUN)
-						.semiPierce(2, getPrimaryDamage())
-						.explosion(new ServerHitscanHandler.HitscanExplosionData(2f, 0f, 0f, true))
-						.alternate().perform();
+			{
+				ServerHitscanHandler.Hitscan scan = ServerHitscanHandler.makeBasicHitscan(user, ServerHitscanHandler.SLAB, 2 * getPrimaryDamage(), DamageSources.GUN)
+															.semiPierce(2, getPrimaryDamage())
+															.explosion(new ServerHitscanHandler.HitscanExplosionData(2f, 0f, 0f, true))
+															.alternate();
+				if(this instanceof MarksmanRevolverItem)
+					scan = scan.resetHammers();
+				scan.perform();
+			}
 			else
 				ServerHitscanHandler.performHitscan(user, ServerHitscanHandler.NORMAL, getPrimaryDamage());
 			cdm.setCooldown(this, getPrimaryCooldown(), GunCooldownManager.PRIMARY);
@@ -159,6 +164,13 @@ public abstract class AbstractRevolverItem extends AbstractWeaponItem implements
 	protected String getHammerId()
 	{
 		return null;
+	}
+	
+	public void resetAllHammers(ItemStack stack)
+	{
+		setNbt(stack, "hammer1", 1);
+		setNbt(stack, "hammer2", 1);
+		setNbt(stack, "hammer3", 1);
 	}
 	
 	@Override